+++ .idea/sticky.iml
@@ -6,6 +6,7 @@
       <excludeFolder url="file://$MODULE_DIR$/docs/.contented" />
       <excludeFolder url="file://$MODULE_DIR$/.vscode" />
       <excludeFolder url="file://$MODULE_DIR$/pnpm-lock.yaml" />
+      <excludeFolder url="file://$MODULE_DIR$/packages/docs/.contented" />
       <excludePattern pattern="*.turbo*" />
       <excludePattern pattern="*dist*" />
     </content>
