--- packages/sticky-testcontainers/package.json
@@ -1,48 +1,0 @@
-{
-  "name": "@birthdayresearch/sticky-testcontainers",
-  "version": "0.0.0",
-  "repository": {
-    "url": "git+https://github.com/BirthdayResearch/sticky"
-  },
-  "main": "./dist/index.js",
-  "source": "./src/index.ts",
-  "types": "./dist/index.d.ts",
-  "files": [
-    "dist/**"
-  ],
-  "scripts": {
-    "build": "tsc -b ./tsconfig.build.json",
-    "clean": "rm -rf dist",
-    "lint": "eslint .",
-    "test": "jest"
-  },
-  "lint-staged": {
-    "*": [
-      "prettier --write --ignore-unknown"
-    ],
-    "*.{js,jsx,ts,tsx}": [
-      "eslint --fix",
-      "prettier --write"
-    ]
-  },
-  "eslintConfig": {
-    "parserOptions": {
-      "project": "./tsconfig.json"
-    },
-    "extends": [
-      "@birthdayresearch"
-    ]
-  },
-  "jest": {
-    "displayName": "test",
-    "preset": "@birthdayresearch/sticky-turbo-jest"
-  },
-  "dependencies": {
-    "testcontainers": "^9.8.0"
-  },
-  "devDependencies": {
-    "@birthdayresearch/sticky-turbo-jest": "workspace:*",
-    "@types/node-fetch": "2.6.2",
-    "node-fetch": "2.6.7"
-  }
-}
