--- packages/sticky-testcontainers/tsconfig.json
@@ -1,6 +1,0 @@
-{
-  "extends": "@birthdayresearch/sticky-typescript",
-  "compilerOptions": {
-    "rootDir": "./"
-  }
-}
