--- packages/sticky-turbo-jest/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "@birthdayresearch/sticky-turbo-jest",
-  "version": "0.0.0",
-  "repository": {
-    "url": "git+https://github.com/BirthdayResearch/sticky"
-  },
-  "files": [
-    "jest-preset.js",
-    "jest-turbo.js"
-  ],
-  "dependencies": {
-    "@birthdayresearch/sticky-jest": "workspace:*",
-    "@birthdayresearch/sticky-turbo": "workspace:*"
-  }
-}
