+++ packages/testcontainers/tsconfig.json
@@ -1,0 +1,6 @@
+{
+  "extends": "@stickyjs/typescript",
+  "compilerOptions": {
+    "rootDir": "./"
+  }
+}
