+++ packages/turbo-jest/README.md
@@ -1,11 +1,11 @@
 # Turbo Jest
 
-`@birthdayresearch/sticky-turbo-jest`
+`@stickyjs/turbo-jest`
 
 Turbo aware `jest-preset`; automatically run `dependsOn` script before running jest. By taking advantage of
 content-aware hashing from turborepo. `dependsOn` only run if the pipeline `inputs` have changed.
 
-When you run your test with `@birthdayresearch/sticky-turbo-jest` preset, it uses `displayName` from your jest config to
+When you run your test with `@stickyjs/turbo-jest` preset, it uses `displayName` from your jest config to
 figure out which `pipeline` to use. It will automatically pinpoint the corresponding `dependsOn` and run all the scripts
 specified within.
 
@@ -28,11 +28,11 @@ E.g. `turbo.json` and `package.json`:
     "build:docker": "docker build ..."
   },
   "jest": {
-    "preset": "@birthdayresearch/sticky-turbo-jest",
+    "preset": "@stickyjs/turbo-jest",
     "projects": [
       {
         "displayName": "test:e2e",
-        "preset": "@birthdayresearch/sticky-turbo-jest",
+        "preset": "@stickyjs/turbo-jest",
         "testRegex": ".*\\.e2e\\.ts$"
       }
     ]
