+++ packages/turbo/package.json
@@ -1,15 +1,14 @@
 {
-  "name": "@birthdayresearch/sticky-jest",
+  "name": "@stickyjs/turbo",
   "version": "0.0.0",
   "repository": {
-    "url": "git+https://github.com/BirthdayResearch/sticky"
+    "url": "git+https://github.com/levaintech/sticky"
   },
   "main": "./dist/index.js",
   "source": "./src/index.ts",
   "types": "./dist/index.d.ts",
   "files": [
-    "dist/**",
-    "jest-preset.js"
+    "dist/**"
   ],
   "scripts": {
     "build": "tsc -b ./tsconfig.build.json",
@@ -17,23 +16,30 @@
     "lint": "eslint .",
     "test": "jest"
   },
+  "lint-staged": {
+    "*": [
+      "prettier --write --ignore-unknown"
+    ],
+    "*.{js,jsx,ts,tsx}": [
+      "eslint --fix",
+      "prettier --write"
+    ]
+  },
   "eslintConfig": {
     "parserOptions": {
       "project": "./tsconfig.json"
     },
     "extends": [
-      "@birthdayresearch"
+      "@stickyjs"
     ]
   },
   "jest": {
-    "preset": "./jest-preset.js"
+    "preset": "@stickyjs/jest"
   },
   "dependencies": {
-    "@types/jest": "^29.5.2",
-    "jest": "29.5.0",
-    "jest-extended": "4.0.0",
-    "ts-jest": "29.1.0",
-    "wait-for-expect": "^3.0.2",
-    "whatwg-fetch": "^3.6.2"
+    "turbo": "1.10.3"
+  },
+  "devDependencies": {
+    "@stickyjs/jest": "workspace:*"
   }
 }
