+++ packages/typescript/README.md
@@ -1,12 +1,12 @@
 # Typescript
 
-`@birthdayresearch/sticky-typescript`
+`@stickyjs/typescript`
 
 Within your `tsconfig.json`:
 
 ```json
 {
-  "extends": "@birthdayresearch/sticky-typescript/tsconfig.json",
+  "extends": "@stickyjs/typescript/tsconfig.json",
   "compilerOptions": {
     "rootDir": "./src",
     "outDir": "./dist"
