+++ packages/typescript/package.json
@@ -1,8 +1,8 @@
 {
-  "name": "@birthdayresearch/sticky-typescript",
+  "name": "@stickyjs/typescript",
   "version": "0.0.0",
   "repository": {
-    "url": "git+https://github.com/BirthdayResearch/sticky"
+    "url": "git+https://github.com/levaintech/sticky"
   },
   "files": [
     "tsconfig.json"
